+++ frontend/@/components/ui/animations/Starfield.tsx
@@ -3,7 +3,7 @@
 
 import { useMemo } from "react";
 
-const secureRandom = ()=> {
+export const secureRandom = ()=> {
   const arr = new Uint32Array(1);
   window.crypto.getRandomValues(arr);
   return arr[0] / (0xffffffff + 1);
